# perf(matrices): make randMatrix faster for sparse matrices