--- a/sympy/matrices/dense.py
+++ b/sympy/matrices/dense.py
@@ -1183,32 +1183,31 @@ def randMatrix(r, c=None, min=0, max=99, seed=None, symmetric=False,
     [70,  0,  0],
     [ 0,  0, 88]
     """
-    if c is None:
-        c = r
     # Note that ``Random()`` is equivalent to ``Random(None)``
     prng = prng or random.Random(seed)
 
-    if not symmetric:
-        m = Matrix._new(r, c, lambda i, j: prng.randint(min, max))
-        if percent == 100:
-            return m
-        z = int(r*c*(100 - percent) // 100)
-        m._mat[:z] = [S.Zero]*z
-        prng.shuffle(m._mat)
+    if c is None:
+        c = r
 
-        return m
-
-    # Symmetric case
-    if r != c:
+    if symmetric and r != c:
         raise ValueError('For symmetric matrices, r must equal c, but %i != %i' % (r, c))
-    m = zeros(r)
-    ij = [(i, j) for i in range(r) for j in range(i, r)]
+
+    ij = range(r * c)
     if percent != 100:
         ij = prng.sample(ij, int(len(ij)*percent // 100))
 
-    for i, j in ij:
-        value = prng.randint(min, max)
-        m[i, j] = m[j, i] = value
+    m = zeros(r, c)
+
+    if not symmetric:
+        for ijk in ij:
+            i, j = divmod(ijk, c)
+            m[i, j] = prng.randint(min, max)
+    else:
+        for ijk in ij:
+            i, j = divmod(ijk, c)
+            if i <= j:
+                m[i, j] = m[j, i] = prng.randint(min, max)
+
     return m