<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->


#### Brief description of what is fixed or changed

Makes randMatrix faster for sparse matrices e.g.:
```python
In [1]: %time ok = randMatrix(1000, percent=1)
CPU times: user 139 ms, sys: 5.56 ms, total: 145 ms
Wall time: 144 ms
```
On master it is much slower:
```python
In [1]: %time ok = randMatrix(1000, percent=1)
CPU times: user 6.69 s, sys: 48 ms, total: 6.74 s
Wall time: 6.79 s
```
Also  a slow test was marked as `@slow` because it takes ~3 minutes

#### Other comments


#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
NO ENTRY
<!-- END RELEASE NOTES -->