@smichr 2021-05-22
I notice that this (and the old version) are not assured to fill to the desired percent:
```python
>>> len([i for i in randMatrix(10,symmetric=1,percent=50,min=1) if i])
56
>>> len([i for i in randMatrix(10,symmetric=1,percent=50,min=1) if i])
44
```
Note that the min was 1 so non-zero values should have been mapped into 50 slots.

We need to map half of the desired percentage into the lower half of the symmetric matrix. This isn't a mod c calculation, however.