@smichr 2021-05-22
> This isn't a mod c calculation, however.

about half of the selected indices will have r <= c and then these are duplicated to the upper half of the matrix. I either case (symmetric or not) this is only an approximate filling of the matrix but for a random matrix this should be good enough.